Update README
Updated the paths for bits to point to the correct files. 
Added a section to describe how to utilize the downsampled set instead, along with the LIBMOLGRID support section. 
Updated the data structure section with the ligand PDB, the uff files, as well as the gninatypes files.
Changed the types file naming convention section to match what was used in the cnnaffinitypaper repository.